+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Added `\Wizaplace\SDK\Catalog\CompanySummary::isProfessional`
 - Added `\Wizaplace\SDK\Catalog\CompanySummary::getAverageRating`
+- Added `\Wizaplace\SDK\MailingList\MailingListService::isSubscribed`
 
 ### Bugfixes